finish azure upload provider

+++ backuper/upload_providers/azure.py
@@ -35,15 +35,17 @@ class UploadProviderAzure(
     def _post_save(self, backup_file: Path) -> str:
         zip_backup_file = core.run_create_zip_archive(backup_file=backup_file)
 
-        backup_dest_in_bucket = "{}/{}".format(
+        backup_dest_in_azure_container = "{}/{}".format(
             zip_backup_file.parent.name,
             zip_backup_file.name,
         )
         blob_client = self.blob_service_client.get_blob_client(
-            container=self.container_name, blob=backup_dest_in_bucket
+            container=self.container_name, blob=backup_dest_in_azure_container
         )
 
-        log.info("start uploading %s to %s", zip_backup_file, backup_dest_in_bucket)
+        log.info(
+            "start uploading %s to %s", zip_backup_file, backup_dest_in_azure_container
+        )
 
         with open(file=zip_backup_file, mode="rb") as data:
             blob_client.upload_blob(data=data)
@@ -51,12 +53,14 @@ class UploadProviderAzure(
         log.info(
             "uploaded %s to %s in %s",
             zip_backup_file,
-            backup_dest_in_bucket,
+            backup_dest_in_azure_container,
             self.container_name,
         )
-        return backup_dest_in_bucket
+        return backup_dest_in_azure_container
 
-    def _clean(self, backup_file: Path, max_backups: int) -> None:
+    def _clean(
+        self, backup_file: Path, max_backups: int, min_retention_days: int
+    ) -> None:
         for backup_path in backup_file.parent.iterdir():
             core.remove_path(backup_path)
             log.info("removed %s from local disk", backup_path)
@@ -72,5 +76,17 @@ class UploadProviderAzure(
 
         while len(backup_list_cloud) > max_backups:
             backup_to_remove = backup_list_cloud.pop()
+            file_name = backup_to_remove.split("/")[-1]
+            if core.file_before_retention_period_ends(
+                backup_name=file_name, min_retention_days=min_retention_days
+            ):
+                log.info(
+                    "there are more backups than max_backups (%s/%s), "
+                    "but oldest cannot be removed due to min retention days",
+                    len(backup_list_cloud),
+                    max_backups,
+                )
+                break
+
             self.container_client.delete_blob(blob=backup_to_remove)
             log.info("deleted backup %s from azure blob storage", backup_to_remove)
